Transform Consolidation The old Position state was split into Transform and Location. TransformState will be used to keep track of position data. LocationState created to hold the platform details about where the entity is "located".
Updated Account Details Logic
This should simplify the logic used to get the Account Details.
The account details and the zone details attached to the account is split, so the logic was updated to better flow,
Removed a recursive function that was used to find the account and zone details.
With the updates to the logic I simplified the repositories for the data of these areas.
They will not throw exception, they will use default/null and the logic will now check for IsFound() on the entity.
I think this better helps with the flow.
The Exception were no removed and will now be used by the application to bubble up extreme errors.

Event/Command/Handler Cleanup
Fixed areas that were not trigging updates for Players when properties were changed.
Updated events and commands to use constructor creation, instead of property setting of data.

General ASP.NET Core 3.1 migration updates.
Removed Newtonsoft.Json references.