Capture Editor Log Messages Added SignalR Hub to capture messages from Platform Connected Clients. Clients as in Browser requested clients.
Created a Logging Provider for .NET.
Connection to a new SignalR Hub for capturing Logs on Clients.

Added a background process that will send log messages from the client to the backend server.
Works in batches of 1000 every 100ms.

Created an event that triggers when the AccessToken is set.
Added EnrichWith for the AuthenticatedUserId and DeploymentDetails.UserId on Valid Login.

Platform Monitor
Added InterfaceId for Client/Editor tracking.

General Formatting and Styles.

C81B4FA2-DF52-4493-BD5E-45D757426743